requirements.txt
asgiref==3.8.1
certifi==2024.7.4
charset-normalizer==3.3.2
Django==4.2.11
django-ckeditor==6.7.1
django-jazzmin==3.0.0
django-js-asset==2.2.0
django-redis-cache==3.0.1
idna==3.7
pillow==10.3.0
pyTelegramBotAPI==4.20.0
python-decouple==3.8
redis==3.5.3
requests==2.32.3
sqlparse==0.5.0
typing_extensions==4.12.2
urllib3==2.2.2
